refactoring: moved helpers function from main.py to tools.py
